Rename qiskit.compiler inner modules to avoid name conflict
The qiskit.compiler module has a name conflict with the re-exported
path to public functions qiskit.compiler. Whether python uses
the module or the function is an import time race condition, amd
could be triggered by subtle changes in import evaluation order. To make
the imports deterministic either the function or the module needs to be
renamed and since the module is less commonly used than the function
that is renamed.
