+++ qiskit/compiler/assembler.py
