+++ qiskit/compiler/scheduler.py
