+++ qiskit/compiler/sequencer.py
