+++ test/python/compiler/test_assembler.py
@@ -21,7 +21,7 @@ import numpy as np
 import qiskit.pulse as pulse
 from qiskit.circuit import Instruction, Gate, Parameter, ParameterVector, ParameterExpression
 from qiskit.circuit import QuantumRegister, ClassicalRegister, QuantumCircuit
-from qiskit.compiler.assemble import assemble
+from qiskit.compiler.assembler import assemble
 from qiskit.exceptions import QiskitError
 from qiskit.pulse import Schedule, Acquire, Play
 from qiskit.pulse.channels import MemorySlot, AcquireChannel, DriveChannel, MeasureChannel
